chore: Move dynamic TLS handler to common package
